move to timelib.WithTimeout
