+++ lib/log/log.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"runtime/debug"
 	"testing"
-	"time"
 
 	"cdr.dev/slog"
 	"cdr.dev/slog/sloggers/sloghuman"
@@ -112,17 +111,4 @@ func Stderr(ctx context.Context) context.Context {
 	stdlog.SetOutput(sl.Writer())
 
 	return With(ctx, l)
-}
-
-// WithTimeout returns context.WithTimeout(ctx, timeout) but timeout is overridden with D2_TIMEOUT if set
-func WithTimeout(ctx context.Context, timeout time.Duration) (context.Context, context.CancelFunc) {
-	t := timeout
-	if seconds, has := env.Timeout(); has {
-		t = time.Duration(seconds) * time.Second
-	}
-	if t <= 0 {
-		return ctx, func() {}
-	}
-
-	return context.WithTimeout(ctx, t)
 }
